requirements.txt
bcrypt==5.0.0
cffi==2.0.0
colorama==0.4.6
cryptography==46.0.4
flexcache==0.3
flexparser==0.4
importlib_resources==6.5.2
iniconfig==2.3.0
invoke==2.2.1
msvc_runtime==14.44.35112
numpy==2.4.2
packaging==26.0
paramiko==4.0.0
phoenix6==26.1.1
Pint==0.25.2
platformdirs==4.5.1
pluggy==1.6.0
pycparser==3.0
pyfrc==2026.0.1
Pygments==2.19.2
PyNaCl==1.6.2
pynetconsole==2.0.4
pyntcore==2026.2.1
pytest==9.0.2
pytest-reraise==2.1.2
robotpy==2026.2.1
robotpy-cli==2026.0.1
robotpy-ctre==2026.1.0
robotpy-hal==2026.2.1
robotpy-halsim-gui==2026.2.1
robotpy-installer==2026.0.1
robotpy-native-ntcore==2026.2.1
robotpy-native-wpihal==2026.2.1
robotpy-native-wpilib==2026.2.1
robotpy-native-wpimath==2026.2.1
robotpy-native-wpinet==2026.2.1
robotpy-native-wpiutil==2026.2.1
robotpy-navx==2026.0.0
robotpy-rev==2026.0.1
robotpy-wpilib-utilities==2026.0.1
robotpy-wpimath==2026.2.1
robotpy-wpinet==2026.2.1
robotpy-wpiutil==2026.2.1
sleipnirgroup-choreolib==2026.0.1
tomli==2.4.0
tomlkit==0.14.0
typing_extensions==4.15.0
wpilib==2026.2.1
